Update broker renewer APIs and logging
This commit updates the broker to always log errors (which helps track
down issues when debugging) and switches to the new renewer API that is
bundled with Vault's golang library.

The previous implementation of the renewer relied on timers and
carefully updating binding information. If the broker crashed and data
was not persisted, the broker would either fail to start or fail to
renew tokens. The new API is significantly less fragile to these types
of situations.

One potential downside to the new API is that restoring a lot of binds
could potentially DOS the Vault cluster. While this is rare, it is
something to keep in mind. At book, all persisted leases will attempt
to renew to calculate their remaining time, since we don't know the time
elapsed between when the broker died and its restart. This problem
existed in the old API as well.

- Fixes GH-12
- Fixes GH-6
- Fixes GH-4